requirements.txt
slack_bolt
azure-storage-blob
flake8
black
isort